- increased floating point precision in shaders due to artifacts

+++ src/rendersystems/OpenGL/GLES/OpenGLES_defaultShaders.h
@@ -20,10 +20,10 @@
 #define SHADER_Plain_Include "\
 "
 #define SHADER_Textured_Include "\
-	varying mediump vec2 texFrag; \n\
+	varying highp vec2 texFrag; \n\
 "
 #define SHADER_Colored_Include "\
-	varying mediump vec4 colorFrag; \n\
+	varying highp vec4 colorFrag; \n\
 "
 #define SHADER_ColoredTextured_Include "\
 " SHADER_Colored_Include "\
@@ -37,33 +37,33 @@
 "
 #define SHADER_VERTEX_Plain_Include "\
 " SHADER_VERTEX_Include "\
-	uniform mediump vec4 systemColor; \n\
+	uniform highp vec4 systemColor; \n\
 "
 #define SHADER_VERTEX_Textured_Include "\
 " SHADER_VERTEX_Include "\
 " SHADER_Textured_Include "\
-	uniform mediump vec4 systemColor; \n\
-	attribute mediump vec2 tex; \n\
+	uniform highp vec4 systemColor; \n\
+	attribute highp vec2 tex; \n\
 "
 #define SHADER_VERTEX_Colored_Include "\
 " SHADER_VERTEX_Include "\
 " SHADER_Colored_Include "\
-	attribute mediump vec4 color; \n\
+	attribute highp vec4 color; \n\
 "
 #define SHADER_VERTEX_ColoredTextured_Include "\
 " SHADER_VERTEX_Include "\
 " SHADER_ColoredTextured_Include "\
-	attribute mediump vec4 color; \n\
-	attribute mediump vec2 tex; \n\
+	attribute highp vec4 color; \n\
+	attribute highp vec2 tex; \n\
 "
 
 // pixel header defines
 #define SHADER_PIXEL_Plain_Include "\
-	uniform mediump vec4 systemColor; \n\
+	uniform highp vec4 systemColor; \n\
 "
 #define SHADER_PIXEL_Textured_Include "\
 " SHADER_Textured_Include "\
-	uniform mediump vec4 systemColor; \n\
+	uniform highp vec4 systemColor; \n\
 	uniform sampler2D sampler2d; \n\
 "
 #define SHADER_PIXEL_Colored_Include "\
@@ -77,7 +77,7 @@
 // pixel header defines (ALPHA-HACK)
 #define SHADER_PIXEL_Textured_AlphaHack_Include "\
 " SHADER_Textured_Include "\
-	uniform mediump vec4 systemColor; \n\
+	uniform highp vec4 systemColor; \n\
 	uniform sampler2D sampler2d; \n\
 	uniform sampler2D sampler2dAlpha; \n\
 "
@@ -131,7 +131,7 @@
 	} \n\
 "
 #define SHADER_PixelLerp SHADER_PIXEL_Plain_Include "\
-	uniform mediump float lerpAlpha; \n\
+	uniform highp float lerpAlpha; \n\
 	void main(void) \n\
 	{ \n\
 		gl_FragColor = vec4(mix(vec3(1.0, 1.0, 1.0), systemColor.rgb, lerpAlpha), systemColor.a); \n\
@@ -151,10 +151,10 @@
 	} \n\
 "
 #define SHADER_PixelTexturedLerp SHADER_PIXEL_Textured_Include "\
-	uniform mediump float lerpAlpha; \n\
+	uniform highp float lerpAlpha; \n\
 	void main(void) \n\
 	{ \n\
-		mediump vec4 tex = texture2D(sampler2d, texFrag); \n\
+		highp vec4 tex = texture2D(sampler2d, texFrag); \n\
 		gl_FragColor = vec4(mix(tex.rgb, systemColor.rgb, lerpAlpha), tex.a * systemColor.a); \n\
 	} \n\
 "
@@ -172,7 +172,7 @@
 	} \n\
 "
 #define SHADER_PixelColoredLerp SHADER_PIXEL_Colored_Include "\
-	uniform mediump float lerpAlpha; \n\
+	uniform highp float lerpAlpha; \n\
 	void main(void) \n\
 	{ \n\
 		gl_FragColor = vec4(mix(vec3(1.0, 1.0, 1.0), colorFrag.rgb, lerpAlpha), colorFrag.a); \n\
@@ -192,10 +192,10 @@
 	} \n\
 "
 #define SHADER_PixelColoredTexturedLerp SHADER_PIXEL_ColoredTextured_Include "\
-	uniform mediump float lerpAlpha; \n\
+	uniform highp float lerpAlpha; \n\
 	void main(void) \n\
 	{ \n\
-		mediump vec4 tex = texture2D(sampler2d, texFrag); \n\
+		highp vec4 tex = texture2D(sampler2d, texFrag); \n\
 		gl_FragColor = vec4(mix(tex.rgb, colorFrag.rgb, lerpAlpha), tex.a * colorFrag.a); \n\
 	} \n\
 "
@@ -204,7 +204,7 @@
 #define SHADER_PixelTexturedMultiply_AlphaHack SHADER_PIXEL_Textured_AlphaHack_Include "\
 	void main(void) \n\
 	{ \n\
-		mediump vec4 newColor = vec4(texture2D(sampler2d, texFrag).rgb, texture2D(sampler2dAlpha, texFrag).r) * systemColor; \n\
+		highp vec4 newColor = vec4(texture2D(sampler2d, texFrag).rgb, texture2D(sampler2dAlpha, texFrag).r) * systemColor; \n\
 		if (newColor.a < " SHADER_AlphaHack_ALPHA_THRESHOLD ") \n\
 		{ \n\
 			discard; \n\
@@ -213,10 +213,10 @@
 	} \n\
 "
 #define SHADER_PixelTexturedLerp_AlphaHack SHADER_PIXEL_Textured_AlphaHack_Include "\
-	uniform mediump float lerpAlpha; \n\
+	uniform highp float lerpAlpha; \n\
 	void main(void) \n\
 	{ \n\
-		mediump vec4 newColor = vec4(mix(texture2D(sampler2d, texFrag).rgb, systemColor.rgb, lerpAlpha), texture2D(sampler2dAlpha, texFrag).r * systemColor.a); \n\
+		highp vec4 newColor = vec4(mix(texture2D(sampler2d, texFrag).rgb, systemColor.rgb, lerpAlpha), texture2D(sampler2dAlpha, texFrag).r * systemColor.a); \n\
 		if (newColor.a < " SHADER_AlphaHack_ALPHA_THRESHOLD ") \n\
 		{ \n\
 			discard; \n\
@@ -228,7 +228,7 @@
 #define SHADER_PixelColoredTexturedMultiply_AlphaHack SHADER_PIXEL_ColoredTextured_AlphaHack_Include "\
 	void main(void) \n\
 	{ \n\
-		mediump vec4 newColor = vec4(texture2D(sampler2d, texFrag).rgb, texture2D(sampler2dAlpha, texFrag).r) * colorFrag; \n\
+		highp vec4 newColor = vec4(texture2D(sampler2d, texFrag).rgb, texture2D(sampler2dAlpha, texFrag).r) * colorFrag; \n\
 		if (newColor.a < " SHADER_AlphaHack_ALPHA_THRESHOLD ") \n\
 		{ \n\
 			discard; \n\
@@ -237,10 +237,10 @@
 	} \n\
 "
 #define SHADER_PixelColoredTexturedLerp_AlphaHack SHADER_PIXEL_ColoredTextured_AlphaHack_Include "\
-	uniform mediump float lerpAlpha; \n\
+	uniform highp float lerpAlpha; \n\
 	void main(void) \n\
 	{ \n\
-		mediump vec4 newColor = vec4(mix(texture2D(sampler2d, texFrag).rgb, colorFrag.rgb, lerpAlpha), texture2D(sampler2dAlpha, texFrag).r * colorFrag.a); \n\
+		highp vec4 newColor = vec4(mix(texture2D(sampler2d, texFrag).rgb, colorFrag.rgb, lerpAlpha), texture2D(sampler2dAlpha, texFrag).r * colorFrag.a); \n\
 		if (newColor.a < " SHADER_AlphaHack_ALPHA_THRESHOLD ") \n\
 		{ \n\
 			discard; \n\
